+++ app/src/main/res/values/strings.xml
@@ -1,8 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <string name="app_name">My App</string>
-
-    <string name="author_format">Author: %s</string>
-
-    <string name="default_base_url">https://api.github.com</string>
+    <string name="default_base_url"></string>
 </resources>
