+++ app/src/sample/java/com/mycompany/myapp/app/VariantApplicationComponent.java
@@ -1,0 +1,9 @@
+package com.mycompany.myapp.app;
+
+
+import com.mycompany.myapp.ui.main.SampleComponent;
+import com.mycompany.myapp.ui.main.SampleComponent.SampleModule;
+
+public interface VariantApplicationComponent {
+    SampleComponent sampleComponent(SampleModule module);
+}
