+++ app/src/sample/java/com/mycompany/myapp/data/api/github/model/Commit.java
