+++ app/src/sample/res/layout/activity_sample.xml
@@ -21,6 +21,6 @@
 
     </android.support.design.widget.AppBarLayout>
 
-    <include layout="@layout/content_main"/>
+    <include layout="@layout/content_sample"/>
 
 </android.support.design.widget.CoordinatorLayout>
