+++ app/src/sample/res/layout/content_sample.xml
