+++ app/src/sample/res/layout/fragment_sample.xml
