+++ app/src/sample/res/layout/row_commit_summary.xml
