+++ app/src/sample/res/values/strings.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="author_format">Author: %s</string>
+
+    <string name="default_base_url">https://api.github.com</string>
+</resources>
